+++ src/MeisterProPR.Application/Interfaces/IThreadMemoryService.cs
@@ -38,7 +38,7 @@ public interface IThreadMemoryService
         Guid clientId,
         string repositoryId,
         int pullRequestId,
-        int threadId,
+        long threadId,
         string? previousStatus,
         string currentStatus,
         string reason,
