+++ src/MeisterProPR.Application/Services/ReviewOrchestrationService.LogMessages.cs
@@ -46,10 +46,10 @@ public sealed partial class ReviewOrchestrationService
     private static partial void LogReviewCompleted(ILogger logger, Guid jobId);
 
     [LoggerMessage(Level = LogLevel.Information, Message = "Thread {ThreadId} on PR#{PrId} marked as fixed")]
-    private static partial void LogThreadResolved(ILogger logger, int threadId, int prId);
+    private static partial void LogThreadResolved(ILogger logger, long threadId, int prId);
 
     [LoggerMessage(Level = LogLevel.Warning, Message = "Thread {ThreadId} evaluation failed on PR#{PrId} — skipping")]
-    private static partial void LogThreadEvaluationFailed(ILogger logger, int threadId, int prId, Exception ex);
+    private static partial void LogThreadEvaluationFailed(ILogger logger, long threadId, int prId, Exception ex);
 
     [LoggerMessage(
         Level = LogLevel.Warning,
